requirements.txt
telethon>=1.36.0
httpx>=0.27.0
python-dotenv>=1.0.1
uvloop>=0.20.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.2
ruff>=0.5.5
mypy>=1.10.0
black>=24.8.0
pytest>=8.2.0
